requirements.txt
rich
dnspython
python-nmap
termcolor
python-whois
python-dotenv
requests
zeroconf
trufflehog3